feat(tests): add base level of unit tests

+++ src/utils.py
@@ -16,15 +16,20 @@ def write_json(data, output_file):
         output.write(json_string)
 
 
-def delete_file(target_file):
+def delete_file(target_file: str):
+    """
+    helper to attempt to delete a file.
+    :param target_file: location of file to delete
+    :return:
+    """
     logging.info(f"Deleting {target_file}")
-    os.remove(target_file)
+    if os.path.exists(target_file):
+        os.remove(target_file)
 
 
 def create_directory_if_not_exists(directory: str):
     """
-    create_directory_if_not_exists will
-    :param directory:
+    :param directory: directory to create if it does not already exist
     :return:
     """
     logging.info(f"Checking to see if {directory} exists")
@@ -33,17 +38,16 @@ def create_directory_if_not_exists(directory: str):
         os.mkdir(directory)
 
 
-def write_csv(data, field_names, output_file):
+def write_csv(data, output_file):
     """
-    write_csv will take a given list and write it out to an output file.
-    :param data:
-    :param field_names:
-    :param output_file:
+    write_csv will take a given list and write it out to an output file. Each row object needs to implement a `to_csv`
+    method.
+    :param data: list of objects with a to_csv method
+    :param output_file: location of file to write data to
     :return:
     """
     with open(output_file, mode="w") as csv_file:
-        writer = csv.DictWriter(csv_file, field_names)
-        writer.writeheader()
+        writer = csv.writer(csv_file, delimiter=",")
         for row in data:
             # TODO: Not particularly happy about this being hardcoded. This logic should live in the Error class
-            writer.writerow({"LINE_NUM": row.row, "ERROR_MSG": row.message})
+            writer.writerow(row.to_csv())
